Move popup, background and help to src/
