+++ src/popup/popup.scss
