+++ src/popup/popup.tsx
@@ -4,7 +4,7 @@ import BrowserPopup from './BrowserPopup';
 // import Semantic-ui packages
 import 'semantic-ui-css/semantic.min.css';
 import './popup.scss';
-import '../../../assets/icon.png';
+import '../../assets/icon.png';
 
 console.log('Przypis Powszechny popup script working!');
