+++ src/popup/toggle/toggle.scss
